requirements.txt
pandas>=1.3.0
pyarrow>=5.0.0
PyYAML>=5.4.0
numpy>=1.20.0 